+++ src/main/java/um/tesoreria/core/kotlin/model/view/ChequeraCuotaDeuda.kt
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat
 import com.fasterxml.jackson.annotation.JsonIgnore
 import jakarta.persistence.*
 import org.hibernate.annotations.Immutable
+import um.tesoreria.core.hexagonal.chequeraSerie.infrastructure.persistence.entity.ChequeraSerieEntity
 import um.tesoreria.core.kotlin.model.*
 import java.math.BigDecimal
 import java.time.OffsetDateTime
@@ -102,7 +103,7 @@ data class ChequeraCuotaDeuda(
 
     @OneToOne(optional = true)
     @JoinColumn(name = "chequeraId", referencedColumnName = "clave", insertable = false, updatable = false)
-    var chequeraSerie: ChequeraSerie? = null,
+    var chequeraSerie: ChequeraSerieEntity? = null,
 
     @OneToOne
     @JoinColumn(name = "chc_id", insertable = false, updatable = false)