+++ src/main/java/um/tesoreria/core/model/Debito.java
@@ -23,7 +23,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import um.tesoreria.core.kotlin.model.Auditable;
-import um.tesoreria.core.kotlin.model.ChequeraSerie;
+import um.tesoreria.core.hexagonal.chequeraSerie.infrastructure.persistence.entity.ChequeraSerieEntity;
 import um.tesoreria.core.kotlin.model.Producto;
 import um.tesoreria.core.util.Jsonifier;
 
@@ -77,7 +77,7 @@ public class Debito extends Auditable implements Serializable {
 
 	@OneToOne(optional = true)
 	@JoinColumn(name = "chequeraId", referencedColumnName = "clave", insertable = false, updatable = false)
-	private ChequeraSerie chequeraSerie;
+	private ChequeraSerieEntity chequeraSerie;
 
 	@OneToOne(optional = true)
 	@JoinColumn(name = "productoId", referencedColumnName = "pro_id", insertable = false, updatable = false)